Increase size of some `struct fcb` members
Component: Flash Circular Buffer

Change the type of sector_cnt and scratch_cnt from a uint8_t to
a uint16_t, allowing for counts > 255.

+++ fs/fcb/include/fcb/fcb.h
@@ -53,8 +53,8 @@ struct fcb {
     /* Caller of fcb_init fills this in */
     uint32_t f_magic;		/* As placed on the disk */
     uint8_t f_version;  	/* Current version number of the data */
-    uint8_t f_sector_cnt;	/* Number of elements in sector array */
-    uint8_t f_scratch_cnt;	/* How many sectors should be kept empty */
+    uint16_t f_sector_cnt;	/* Number of elements in sector array */
+    uint16_t f_scratch_cnt;	/* How many sectors should be kept empty */
     struct flash_area *f_sectors; /* Array of sectors, must be contiguous */
 
     /* Flash circular buffer internal state */
